feat(cli): introduce AI-powered API specification suggestion feature
- Added a new command `Suggest` to the CLI for generating OpenAPI specs or MockForge configs using AI.
- Implemented options for input via file or description, output format selection, and additional endpoint suggestions.
- Enhanced the core functionality with a new `handle_suggest` function to process AI-driven suggestions based on user input.
- Updated dependencies in `Cargo.toml` to include `async-trait` for async trait support.

These changes significantly enhance the CLI's capabilities, allowing users to leverage AI for more efficient API specification generation.

+++ crates/mockforge-core/src/intelligent_behavior/config.rs
@@ -195,7 +195,7 @@ fn default_cache_ttl() -> u64 {
     300 // 5 minutes
 }
 
-fn default_max_history() -> u64 {
+fn default_max_history() -> usize {
     50
 }
